Correct a bug to deal with non-integers values in table L
Homogenity/correlation ratio measures require to have at least 23 individuals per class to be computed (computing a variance requires at leats 2 indiciduals). When dealing with non-integres values in L (e.g. cover, proportion), we must count the number of non-null cells (correpondence) but not the total abundance (which can be less than 1 and thus induces a bug in the if condition)